requirements.txt
python-pptx>=0.6.23
pdfplumber>=0.10.0
